Add test to show a global Flow type
Summary:
This test documents what happens if you use Flow based schema gen with a global Flow scalar. I'll fix this in the next diff in the cases of global custom scalars.

I also add a custom scalar into the schema as it changs a few test snapshots. This diff isolates these changes from the other, more complex, changes in the next diff.

Reviewed By: captbaritone

Differential Revision: D59477643

fbshipit-source-id: bdd8cc9ee095a97954059f35d172f4188e391570