Fix Relay incremental build for enums
Summary: This diff removes an optimization we added that skips rebuilding artifacts for operations with enum changes where the enums are not inlined. We use the project config to tell if enums are inlined or not. This is normally okay, except that in test operations, we always inline enums. This means that incremental builds are missed on test operations. Since this is not an obvious fix, I'm removing this operation until we can find a solution.

Reviewed By: tyao1

Differential Revision: D59943056

fbshipit-source-id: d1e66652879c05b503ad176b073d30610c8b1038