Re-organize hierarchy of docs interally
Summary: This re-organizes the docs sidebar to be a bit more straightforward. It groups the guides into 7 folders that align with what a developer might want to do. This diff just organizes our internal pages for now. The external pages should use a very similar structure but I'll go through them and do it in a follow up diff.

Reviewed By: lynnshaoyu

Differential Revision: D60267884

fbshipit-source-id: 6f114e192747b3233d4fbcdc70640ddfdcb670a2